tests/mse_mp4/1342913/test_ChangeWhileWaitingOnMissingData_mp4.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: resume from waiting even after format change occurred</title>
  <script type="text/javascript" src="../common.js"></script>
  <script type="text/javascript" src="../mediasource.js"></script>
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, 'sourceopen').then(function() {
    ok(true, "Receive a sourceopen event");
    var sb = ms.addSourceBuffer("video/mp4");
    fetchAndLoad(sb, 'v.1280x720.', ['init'], '.mp4')
    .then(fetchAndLoad.bind(null, sb, 'v.1280x720.1.', range(3, 4), '.m4s'))
    .then(fetchAndLoad.bind(null, sb, 'v.1280x720.2.', range(3, 5), '.m4s'))
    .then(function() {
      el.play();
      // let seek to the last audio frame.
      el.currentTime = 6;
      return Promise.all([once(el, 'seeked'), once(el, 'stalled')]);
    })
    .then(function() {
      info("seek completed");
      return fetchAndLoad(sb, 'v.640x300.', range(1,3), '.mp4');
    })
    .then(function() {
      ms.endOfStream();
      return once(el, 'ended');
    }).then(function() {
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
